Fix is_prime function

+++ c/utils.h
@@ -1,6 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-int is_prime(int n); // Test if int is prime
+int is_prime(long n); // Test if int is prime
 long max_prime_factor(long n); // Find max prime factor
+
 #endif /* UTILS_H */
